+++ src/core/types.ts
@@ -2,11 +2,11 @@
  * Flash Buffer types
  *
  * @developer           Elijah Rastorguev
- * @version             1.1.0
- * @build               1004
+ * @version             1.1.5
+ * @build               1006
  * @git                 https://github.com/devsdaddy/flash-buffer/
  * @docs                https://github.com/devsdaddy/flash-buffer/#readme
- * @updated             13.04.2026
+ * @updated             20.04.2026
  */
 import {FlashBuffer} from "./buffer";
 
@@ -34,4 +34,9 @@ export interface Freezable {
 /**
  * Read only binary buffer
  */
-export type ReadonlyFlashBuffer = Omit<FlashBuffer, keyof Freezable | 'writeInt8' | 'writeUint8' | 'writeInt16' | 'writeUint16' | 'writeInt32' | 'writeUint32' | 'writeBigInt64' | 'writeBigUint64' | 'writeFloat32' | 'writeFloat64' | 'writeString' | 'writeBytes' | 'writeVarInt' | 'writeVarUint' | 'writeCString' | 'writeAligned' | 'writeBits'>;
+export type ReadonlyFlashBuffer = Omit<FlashBuffer, keyof Freezable |
+    'writeInt8' | 'writeUint8' | 'writeInt16' | 'writeUint16' | 'writeInt32' | 'writeUint32' |
+    'writeBigInt64' | 'writeBigUint64' | 'writeFloat32' | 'writeFloat64' | 'writeString' | 'writeBytes' |
+    'writeVarInt' | 'writeVarUint' | 'writeCString' | 'writeAligned' | 'writeBits' |
+    'write' | 'writeTypedArray' | 'writeSFixed64' | 'writeSFixed32' | 'writeVarUint64' |
+    'writeSint32' | 'writeSint64' | 'writeBool'>;